third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-position-grid-001-expected.txt

This is a testharness.js-based test.
[FAIL] .target 4
  assert_equals: \n<div class="target target1-b" data-offset-y="145" data-expected-width="100"></div>\noffsetTop expected 145 but got 45
Harness: the test ran to completion.
